Follow up of https://github.com/tinygrad/tinygrad/pull/1187 to fix CIFAR regression https://github.com/tinygrad/tinygrad/pull/1286 and https://github.com/tinygrad/tinygrad/pull/1287

Identified problem was that the heuristic was ranking best places to split, but did not abort when even the best option was bad. This version now skips splitting if (1) the reduction is a small reduction or (2) we'd be forced to reduce with a bad global dim to stride ratio.

I've only tested this on my M1, **so before merging again it would be good to re-test on other envs**!

---

CIFAR specifically was being held up on a single problematic kernel, `r_4096_3_3_2_2_2_4_2_4_8_2_4_4` which on it's own was causing the large boost in memory consumption and a large slowdown.

New CIFAR numbers on M1 Max:
```
  0 1904.08 ms run, 1883.58 ms python,   20.50 ms CL,    2.94 loss, 0.009521 LR, 0.09 GB used,    613.58 GFLOPS
  1 1352.30 ms run, 1351.52 ms python,    0.78 ms CL,    3.02 loss, 0.008464 LR, 7.87 GB used,    863.93 GFLOPS
  2 1241.88 ms run, 1240.57 ms python,    1.31 ms CL,    5.76 loss, 0.007408 LR, 7.87 GB used,    940.75 GFLOPS
  3 1241.01 ms run, 1239.70 ms python,    1.31 ms CL,    7.17 loss, 0.006351 LR, 7.87 GB used,    941.41 GFLOPS
  4 1241.18 ms run, 1239.94 ms python,    1.24 ms CL,    6.99 loss, 0.005294 LR, 7.87 GB used,    941.28 GFLOPS
  5 1242.83 ms run, 1214.35 ms python,   28.48 ms CL,    4.66 loss, 0.004237 LR, 7.87 GB used,    940.03 GFLOPS
  6 1260.45 ms run, 1231.65 ms python,   28.80 ms CL,    3.61 loss, 0.003180 LR, 7.87 GB used,    926.89 GFLOPS
  7 1289.97 ms run, 1287.83 ms python,    2.14 ms CL,    3.33 loss, 0.002124 LR, 7.87 GB used,    905.68 GFLOPS
  8 1253.34 ms run, 1252.23 ms python,    1.10 ms CL,    3.67 loss, 0.001067 LR, 7.87 GB used,    932.15 GFLOPS
  9 1286.83 ms run, 1248.92 ms python,   37.91 ms CL,    2.99 loss, 0.000010 LR, 7.87 GB used,    907.89 GFLOPS
```