@tinyb0t 2023-07-20
Changes made in `tinygrad/`:
```
------------------------------------------------------------
files                             insertions       deletions
------------------------------------------------------------
tinygrad/lazy.py                          17               1
------------------------------------------------------------
lines added in the tinygrad folder: 16
```