@geohot 2023-07-20
What's the CIFAR baseline on M1?

I think it's worth comparing stable diffusion, llama 7B, and CIFAR train speed. If there's no regressions in any of those I think we are good.