@ae99 2023-07-20
CIFAR Training
Master
```
> STEPS=10 python examples/hlb_cifar10.py
  0 1781.43 ms run, 1780.93 ms python,    0.50 ms CL,    2.94 loss, 0.009521 LR, 0.09 GB used,    655.48 GFLOPS
  1 1370.04 ms run, 1369.11 ms python,    0.94 ms CL,    3.02 loss, 0.008464 LR, 7.57 GB used,    852.30 GFLOPS
  2 1270.77 ms run, 1268.08 ms python,    2.69 ms CL,    5.76 loss, 0.007408 LR, 7.57 GB used,    918.88 GFLOPS
  3 1277.09 ms run, 1275.86 ms python,    1.23 ms CL,    7.17 loss, 0.006351 LR, 7.57 GB used,    914.33 GFLOPS
  4 1271.06 ms run, 1269.82 ms python,    1.24 ms CL,    6.98 loss, 0.005294 LR, 7.57 GB used,    918.67 GFLOPS
  5 1269.75 ms run, 1268.52 ms python,    1.23 ms CL,    4.66 loss, 0.004237 LR, 7.57 GB used,    919.61 GFLOPS
  6 1255.20 ms run, 1254.05 ms python,    1.14 ms CL,    3.62 loss, 0.003180 LR, 7.57 GB used,    930.28 GFLOPS
  7 1275.13 ms run, 1273.80 ms python,    1.33 ms CL,    3.33 loss, 0.002124 LR, 7.57 GB used,    915.73 GFLOPS
  8 1284.34 ms run, 1283.13 ms python,    1.21 ms CL,    3.66 loss, 0.001067 LR, 7.57 GB used,    909.17 GFLOPS
  9 1271.79 ms run, 1270.52 ms python,    1.27 ms CL,    2.99 loss, 0.000010 LR, 7.57 GB used,    918.14 GFLOPS
```
This PR
```
> STEPS=10 python examples/hlb_cifar10.py
  0 1900.81 ms run, 1875.45 ms python,   25.36 ms CL,    2.94 loss, 0.009521 LR, 0.09 GB used,    614.64 GFLOPS
  1 1311.74 ms run, 1310.85 ms python,    0.89 ms CL,    3.02 loss, 0.008464 LR, 7.87 GB used,    890.65 GFLOPS
  2 1227.10 ms run, 1225.78 ms python,    1.32 ms CL,    5.76 loss, 0.007408 LR, 7.87 GB used,    952.08 GFLOPS
  3 1231.38 ms run, 1230.00 ms python,    1.39 ms CL,    7.17 loss, 0.006351 LR, 7.87 GB used,    948.77 GFLOPS
  4 1242.24 ms run, 1240.94 ms python,    1.31 ms CL,    6.99 loss, 0.005294 LR, 7.87 GB used,    940.47 GFLOPS
  5 1232.39 ms run, 1212.33 ms python,   20.06 ms CL,    4.66 loss, 0.004237 LR, 7.87 GB used,    948.00 GFLOPS
  6 1253.18 ms run, 1227.26 ms python,   25.92 ms CL,    3.61 loss, 0.003180 LR, 7.87 GB used,    932.27 GFLOPS
  7 1222.61 ms run, 1220.69 ms python,    1.92 ms CL,    3.33 loss, 0.002124 LR, 7.87 GB used,    955.57 GFLOPS
  8 1232.50 ms run, 1224.33 ms python,    8.16 ms CL,    3.67 loss, 0.001067 LR, 7.87 GB used,    947.91 GFLOPS
  9 1269.98 ms run, 1249.66 ms python,   20.33 ms CL,    2.99 loss, 0.000010 LR, 7.87 GB used,    919.93 GFLOPS
```