@ae99 2023-07-20
Getting a small regression now on Stable Diffusion :confused:

This feels like whack-a-mole, going to need to create some sort of fuzzing with random inputs as this change is obviously more brittle than I've been able to manually verify.

Converting to draft.