@geohot 2023-07-20
I'm making some real world tests, runs llama+stable diffusion+cifar training.

https://github.com/tinygrad/tinygrad/pull/1297